# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MS Elevate LEAPS Tracker API
  version: 1.0.0
  description: |
    Complete API documentation for the Microsoft Elevate Indonesia LEAPS Tracker platform.
    
    The LEAPS framework consists of 5 stages:
    - **Learn**: Complete Microsoft Learn courses (20 points)
    - **Explore**: Apply AI in classroom with evidence (50 points)  
    - **Amplify**: Train peers/students (2 points/peer, 1 point/student)
    - **Present**: LinkedIn post with screenshot (20 points)
    - **Shine**: Innovation ideas for recognition
    
    ## Authentication
    All authenticated endpoints require a valid Clerk JWT token in the Authorization header.
    
    ## Rate Limits
    - File uploads: 10MB max per file
    - Amplify submissions: 50 peers, 200 students per 7-day rolling window
    - Bulk operations: Maximum 50 items per request
    
  contact:
    name: MS Elevate Team
    url: https://leaps.mereka.org
servers:
  - url: https://leaps.mereka.org/api
    description: Production server
  - url: https://staging-leaps.mereka.org/api 
    description: Staging server
security:
  - ClerkJWT: []
components:
  securitySchemes:
    ClerkJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk authentication JWT token
  schemas:
...
    Badge:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        icon_url:
          type: string
        criteria:
          type: object
        created_at:
          type: string
          format: date-time
      required:
        - id
        - code
        - name
...
paths:
...
  /admin/exports:
    get:
      summary: Export data to CSV
      tags: [Admin]
      security:
        - ClerkJWT: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [users, submissions, leaderboard, points]
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, xlsx]
            default: csv
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: CSV/XLSX file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /admin/badges:
    get:
      summary: List all badge definitions
      tags: [Admin]
      security:
        - ClerkJWT: []
      responses:
        '200':
          description: List of badges
          content:
            application/json:
              schema:
                type: object
                properties:
                  badges:
                    type: array
                    items:
                      $ref: '#/components/schemas/Badge'

    post:
      summary: Create or update badge definition
      tags: [Admin]
      security:
        - ClerkJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                name:
                  type: string
                description:
                  type: string
                icon_url:
                  type: string
                criteria:
                  type: object
                  description: JSON criteria for badge earning
              required:
                - code
                - name
                - description
                - criteria
      responses:
        '200':
          description: Badge created/updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  badge:
                    $ref: '#/components/schemas/Badge'
...